+++ tests/test_devices.py
@@ -160,6 +160,24 @@ def test_full_flush_contact_interface(fake_home: Home):
     )
 
 
+def test_full_flush_contact_interface6(fake_home: Home):
+    d = fake_home.search_device_by_id("3014F7110000000000056775")
+    assert isinstance(d, FullFlushContactInterface6)
+
+    assert d.functionalChannels[3].binaryBehaviorType == BinaryBehaviorType.NORMALLY_CLOSE
+    assert d.functionalChannels[3].windowState == WindowState.OPEN
+    assert d.functionalChannels[3].multiModeInputMode == MultiModeInputMode.KEY_BEHAVIOR
+
+    assert d.functionalChannels[6].binaryBehaviorType == BinaryBehaviorType.NORMALLY_CLOSE
+    assert d.functionalChannels[6].windowState == None
+    assert d.functionalChannels[6].multiModeInputMode == MultiModeInputMode.KEY_BEHAVIOR
+
+    assert str(d) == (
+        "HmIP-FCI6 BA - Licht (Sensor) lowBat(None) unreach(None) rssiDeviceValue(None) rssiPeerValue(None) "
+        "configPending(False) dutyCycle(None)"
+    )
+
+
 def test_full_flush_input_switch(fake_home: Home):
     d = fake_home.search_device_by_id("3014F7110000000HmIPFSI16")
     assert isinstance(d, FullFlushInputSwitch)
@@ -1289,7 +1307,7 @@ def test_blind_module(fake_home: Home):
             "primaryOpenAdjustable(True) primaryShadingStateType(POSITION_USED) "
             "secondaryCloseAdjustable(False) secondaryOpenAdjustable(False) "
             "secondaryShadingStateType(NOT_EXISTENT) primaryShadingLevel(0.94956) "
-            "secondaryShadingLevel(None) previousPrimaryShadingLevel(None) "
+            "secondaryShadingLevel(0) previousPrimaryShadingLevel(None) "
             "previousSecondaryShadingLevel(None) identifyOemSupported(True) productId(10) "
             "profileMode(AUTOMATIC) userDesiredProfileMode(AUTOMATIC) shadingDriveVersion(None) "
             "shadingPackagePosition(TOP) shadingPositionAdjustmentActive(None) "