+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] 2020-11-09
+### ADDED
+- API
+    - Devices
+      - [HMIP-DRSI4] (Switch Actuator for DIN rail mount – 4x channels)
+      - [HMIP-DRBLI4] (Blind Actuator for DIN rail mount – 4 channels)
+      - [HMIP-FCI6] (Contact Interface flush-mount – 6x channels)
+
 
 ## [0.12.1] 2020-11-10
 ### ADDED